// web/src/shared/api/generated/model/aiServiceStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type AiServiceStatus = typeof AiServiceStatus[keyof typeof AiServiceStatus];


export const AiServiceStatus = {
  NotUsed: 'NotUsed',
  Reachable: 'Reachable',
  Unreachable: 'Unreachable',
} as const;